+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This changelog covers all three packges, as they are (for now) updated as a whol
 ## v0.24.1
 
 - Improve resource selector dropdown, show previews, remove dependency #60
+- Add toast notifications #63
 
 ## v0.24.0